# apps/physics_agent_service/openapi.yaml
openapi: 3.1.0
info:
  title: Physics Agent Service
  description: "The Physics Agent Service takes a USD file (uploaded directly or\
    \ referenced by S3 URI) as input and runs a multi-step VLM-based classification\
    \ pipeline: build_dataset_usd \u2192 prepare_dataset \u2192 predict \u2192 apply_physics.\
    \ It produces predictions (JSONL), an HTML report, and a simulation-ready USD\
    \ with UsdPhysics schemas applied for each prim in the scene.\n\n## Highlights\n\n\
    * End-to-end pipeline\n    * POST a USD file (multipart) or an S3 URI to start the pipeline.\n\
    \    * Two-step upload supported via POST /pipeline/upload-usd followed by POST /pipeline with session_id.\n\
    \    * Produces predictions.jsonl, an HTML classification report, and a simulation-ready\
    \ USD (scene_physics.<output extension>) with UsdPhysics RigidBodyAPI / CollisionAPI\
    \ / MassAPI / MaterialAPI authored on each predicted prim.\n* Flexible\
    \ configuration\n    * Server-side defaults control rendering modes, VLM prompts,\
    \ and prediction settings.\n    * Supports multiple rendering modes (prim_only, composition,\
    \ etc.).\n    * Configurable VLM backend and model.\n    * Selectable rendering\
    \ backend via API parameter:\n        * `remote` (default in the bundled Docker\
    \ Compose) \u2014 HTTP render service. The public compose points this at the OVRTX\
    \ sidecar, so the main service stays CPU-only.\n        * `warp` \u2014 In-process\
...
paths:
  /pipeline/upload-usd:
    post:
      tags:
      - pipeline
      summary: Upload Usd Immediate
      description: Upload a USD file and create a session for later pipeline execution.
      operationId: upload_usd_immediate_pipeline_upload_usd_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_usd_immediate_pipeline_upload_usd_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCreated'
        '403':
          description: Client S3 URI rejected by the configured bucket allowlist, or S3 access denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /pipeline:
    post:
      tags:
      - pipeline
      summary: Create Pipeline
      description: |
        Create and execute a physics agent pipeline.

        At least one input source is required. When more than one is supplied,
        the selected source is deterministic: `session_id` first, then
        `s3_uri`, then `usd_file`. Lower-priority source fields are ignored.
      operationId: create_pipeline_pipeline_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_create_pipeline_pipeline_post'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCreated'
        '403':
          description: Client S3 URI rejected by the configured bucket allowlist, or S3 access denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /pipeline/{session_id}/status:
    get:
      tags:
      - pipeline
      summary: Get Pipeline Status
      description: 'Get pipeline execution status with detailed progress.


        Reads from in-memory event bus state for fast, real-time accuracy.

        Falls back to store-based SessionManager for completed/cross-instance sessions.'
      operationId: get_pipeline_status_pipeline__session_id__status_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineStatus'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /pipeline/{session_id}/results:
    get:
      tags:
      - pipeline
      summary: Get Pipeline Results
      description: Get terminal pipeline execution results.
      operationId: get_pipeline_results_pipeline__session_id__results_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/PipelineResults'
                - $ref: '#/components/schemas/PipelineError'
                title: Response Get Pipeline Results Pipeline  Session Id  Results
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /pipeline/{session_id}/cancel:
    post:
      tags:
      - pipeline
      summary: Cancel Pipeline
      description: 'Cancel a running pipeline.


        Works cross-instance: writes a cancel signal to the store (S3)

        so the executing instance can detect it. Also tries local cancellation.'
      operationId: cancel_pipeline_pipeline__session_id__cancel_post
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /pipeline/{session_id}/events:
    get:
      tags:
      - pipeline
      summary: Stream Progress Events
      description: 'Stream real-time progress events via Server-Sent Events (SSE).


        Only works when connected to the instance running the pipeline.

        For cross-instance progress, use GET /pipeline/{session_id}/status (polling).'
      operationId: stream_progress_events_pipeline__session_id__events_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /pipeline/{session_id}/regenerate:
    post:
      tags:
      - pipeline
      summary: Regenerate Pipeline
      description: Regenerate specific pipeline steps from cached data.
      operationId: regenerate_pipeline_pipeline__session_id__regenerate_post
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegenerateRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCreated'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /pipeline/{session_id}/event-log:
    get:
      tags:
      - pipeline
      summary: Get Event Log
      description: Get the persisted event log for a session.
      operationId: get_event_log_pipeline__session_id__event_log_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /artifacts/{session_id}/predictions:
    get:
      tags:
      - artifacts
      summary: Download Predictions
      description: Download predictions JSONL file.
      operationId: download_predictions_artifacts__session_id__predictions_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /artifacts/{session_id}/report:
    get:
      tags:
      - artifacts
      summary: View Prediction Report
      description: 'View prediction HTML report in browser.


        Generates the report on-demand if it doesn''t exist yet.'
      operationId: view_prediction_report_artifacts__session_id__report_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /sessions/{session_id}:
    get:
      tags:
      - sessions
      summary: Get Session
      description: Get detailed session information.
      operationId: get_session_sessions__session_id__get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - sessions
      summary: Delete Session
      description: Delete a session and all its artifacts.
      operationId: delete_session_sessions__session_id__delete
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_upload_usd_immediate_pipeline_upload_usd_post:
      properties:
        usd_file:
          type: string
          contentMediaType: application/octet-stream
          title: Usd File
          description: USD file to upload (provide this OR s3_uri)
        s3_uri:
          type: string
          title: S3 Uri
          description: S3 URI to a USD file (e.g. s3://bucket/path/scene.usdz)
      type: object
      title: Body_upload_usd_immediate_pipeline_upload_usd_post
...
    ErrorDetail:
      properties:
        detail:
          type: string
          title: Detail
          description: Human-readable error detail
      type: object
      required:
      - detail
      title: ErrorDetail
      description: Standard FastAPI error payload returned by authorization failures.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PipelineStatus:
      properties:
        session_id:
          type: string
          title: Session Id
        status:
          type: string
          title: Status
          description: 'Current status: pending, running, completed, failed, cancelled,
            cancelling'
        current_step:
          anyOf:
          - $ref: '#/components/schemas/CurrentStepInfo'
          - type: 'null'
        completed_steps:
          items:
            $ref: '#/components/schemas/CompletedStepInfo'
          type: array
          title: Completed Steps
        overall_progress:
          $ref: '#/components/schemas/OverallProgress'
        preview_images:
          items:
            type: string
          type: array
          title: Preview Images
          description: URLs to preview images
        can_cancel:
          type: boolean
          title: Can Cancel
          description: Whether pipeline can be cancelled
        elapsed_seconds:
          type: integer
          title: Elapsed Seconds
          description: Total elapsed time in seconds
        created_at:
          type: string
          title: Created At
          description: ISO timestamp when session created
        updated_at:
          type: string
          title: Updated At
          description: ISO timestamp of last update
      type: object
      required:
      - session_id
      - status
      - overall_progress
      - can_cancel
      - elapsed_seconds
      - created_at
      - updated_at
      title: PipelineStatus
      description: Enhanced pipeline execution status with progress.
...
    SessionCreated:
      properties:
        session_id:
          type: string
          title: Session Id
        status:
          type: string
          title: Status
          default: pending
        message:
          type: string
          title: Message
          default: Pipeline queued for execution
        estimated_duration_minutes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Estimated Duration Minutes
          description: Estimated completion time
      type: object
      required:
      - session_id
      title: SessionCreated
      description: Response when session is created.
...